Finished migrating feature-specific components and containers to separate folder.
