--- src/containers/session/SessionContainer.module.css
@@ -1,38 +1,0 @@
-
-.itemWidth {
-  width: 100%;
-}
-
-.sectionMargins {
-  margin-bottom: 1rem;
-}
-
-.containerMargins{
-  margin-bottom: 2rem;
-}
-
-.buttonWidth {
-  width: 90%;
-}
-
-.textSm {
-  font-size: 0.8rem;
-}
-
-.textMd {
-  font-size: 1rem;
-}
-
-.textMb {
-  margin-bottom: 1rem;
-}
-
-.itemBg {
-  &:nth-of-type(even) {
-    background-color: rgba(211, 211, 211, 0.5);
-  }
-}
-
-.itemPadding {
-  padding: 1.5rem 1.5rem 2rem;
-}
