+++ src/features/practice/containers/PracticeContainer.Report.tsx
@@ -1,4 +1,4 @@
-import styles from "./SessionContainer.module.css"
+import styles from "./PracticeContainer.module.css"
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 
@@ -6,8 +6,8 @@ import { usePracticeSessionStore } from "@/src/stores/practiceSessionStore";
 import useQuestionsCorrect from "@/src/hooks/useQuestionsCorrect";
 import useQuestionsAnswered from "@/src/hooks/useQuestionsAnswered";
 
-import SummaryContainer from "containers/summary/SummaryContainer";
-import DetailsContainer from "./SessionContainer.Details";
+import SummaryContainer from "@/src/features/sessionReport/summary/containers/SummaryContainer";
+import DetailsContainer from "../../sessionReport/detail/containers/DetailContainer";
 
 import { ClientStudentResponse } from "@/src/queries/GET/getResponsesBySession";
 import ErrorPage from "@/src/ErrorPage";
@@ -16,8 +16,7 @@ interface SessionContainerProps {
   studentResponses: ClientStudentResponse[]
 }
 
-
-export default function SessionContainer(props: SessionContainerProps) {
+export default function PracticeReportContainer(props: SessionContainerProps) {
 
   const sessionId = usePracticeSessionStore((state) => state.sessionId);
   // const sessionId = usePracticeSessionStore((state) => state.sessionId);
@@ -59,7 +58,13 @@ export default function SessionContainer(props: SessionContainerProps) {
           styles.sectionMargins,
         ].join(" ")}
         onClick={handleShowDetails}
-      >{!showDetails ? "Click to view details +" : "Click to hide details -"}</button>
+      >
+        {
+          !showDetails ?
+            "Click to view details +" :
+            "Click to hide details -"
+        }
+      </button>
       {
         showDetails &&
         <div>
