+++ src/features/sessionReport/detail/containers/DetailContainer.module.css
